move dblp_search function to a seperate file, also add it to the __init__.py
